ci: pipeline test

+++ src/main/java/davi/api/demo/domain/dtos/DomainCreateMoneyDonationConfigDto.java
@@ -1,3 +1,3 @@
 package davi.api.demo.domain.dtos;
 
-public record DomainCreateMoneyDonationConfigDto(boolean enabled, String useDescription, int goal) { }
+public record DomainCreateMoneyDonationConfigDto(boolean enabled, String useDescription, Long goal) { }
