+++ src/main/java/davi/api/demo/domain/entities/MoneyDonationConfig.java
@@ -1,16 +1,16 @@
 package davi.api.demo.domain.entities;
 
 public class MoneyDonationConfig extends DonationConfig {
-    private int goal;
+    private Long goal;
     private String useDescription;
 
     public MoneyDonationConfig(Long id, String uuid, boolean enabled) {
         super(id, uuid, enabled);
     }
 
-    public int getGoal() { return goal; }
+    public Long getGoal() { return goal; }
 
-    public void setGoal(int goal) { this.goal = goal; }
+    public void setGoal(Long goal) { this.goal = goal; }
 
     public String getUseDescription() { return useDescription; }
